# docs/openapi.yaml
openapi: 3.0.3
info:
  title: RSS Reader API
  description: |
    与 Web 前端共用的后端 API。Base URL 请替换为实际部署地址（如 `https://your-host.com/api`）。
    认证：除标注为公开的接口外，均需在请求头携带 `Authorization: Bearer <JWT>`。
  version: 1.0.0

servers:
  - url: https://example.com/api
    description: 替换为实际网关前缀（末尾为 /api）

tags:
  - name: auth
    description: 登录与飞书 OAuth
  - name: categories
    description: 订阅分类
  - name: feeds
    description: 订阅
  - name: articles
    description: 文章列表（订阅内容）
  - name: summary-histories
    description: AI 总结历史

paths:
  /auth/login:
    post:
      tags: [auth]
      summary: 用户名密码登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                  minLength: 6
      responses:
        '200':
          description: 成功，返回 JWT 与用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          description: 账号被锁定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /auth/feishu/login-url:
    get:
      tags: [auth]
      summary: 获取飞书扫码/网页授权入口（公开）
      description: |
        返回本站中转地址 `url` 与飞书 passport 完整授权页 `goto`。
        Web 端通常用 iframe 打开 `goto`；原生 App 可用 WebView/Custom Tabs 打开 `goto`。
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                    example: /api/auth/feishu/login?state=login
                  goto:
                    type: string
                    description: 飞书 OAuth 授权页完整 URL
        '400':
          $ref: '#/components/responses/Error'

  /auth/feishu/callback:
    get:
      tags: [auth]
      summary: 飞书 OAuth 回调（公开）
      description: |
        飞书授权成功后重定向到此地址，携带 `code`、`state`。
        **当前实现返回 HTML**，通过 `window.postMessage` 将 `token` 与 `user` 传给父窗口（Web）。
        原生 Android 若无法使用 postMessage，需配合 WebView 拦截最终 URL、或由服务端另行提供 JSON 换票接口（当前仓库未提供）。
      parameters:
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: state
          schema: { type: string }
      responses:
        '200':
          description: HTML 页面（内含登录结果脚本）

  /auth/feishu/exchange:
    post:
      tags: [auth]
      summary: 使用飞书 code 换取站内 token（JSON）
      description: |
        适用于原生 App 或前后端分离场景。前端在拿到飞书回调中的 code 后，
        调用本接口换取站内 JWT 与用户信息。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                state:
                  type: string
      responses:
        '200':
          description: 成功，返回 JWT 与用户信息
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResult'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          description: 账号被锁定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /feeds:
    get:
      tags: [feeds]
      summary: 订阅列表
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Feed'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /feeds/{id}/refresh:
    post:
      tags: [feeds]
      summary: 立即刷新订阅
      description: 立即抓取当前用户拥有的单个订阅源，并返回刷新后的订阅信息。
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feed'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/Error'
        '502':
          $ref: '#/components/responses/Error'

  /categories:
    get:
      tags: [categories]
      summary: 订阅分类列表
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: 成功返回当前用户的分类列表，按排序号升序
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeedCategory'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [categories]
      summary: 创建订阅分类
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpsertRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedCategory'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 分类名称重复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /categories/reorder:
    put:
      tags: [categories]
      summary: 调整订阅分类排序
      description: 按 `id_list` 顺序重排当前用户的分类，数组越靠前排序越靠前。
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryReorderRequest'
      responses:
        '200':
          description: 排序更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /categories/{id}:
    put:
      tags: [categories]
      summary: 更新订阅分类
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: 分类 ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryUpsertRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeedCategory'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 分类不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '409':
          description: 分类名称重复
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    delete:
      tags: [categories]
      summary: 删除订阅分类
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: 分类 ID
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 分类不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /articles:
    get:
      tags: [articles]
      summary: 文章列表（点击某订阅后传 feed_id 过滤）
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: feed_id
          schema: { type: integer }
          description: 仅查看该订阅下的文章
        - in: query
          name: read
          schema: { type: boolean }
          description: true=仅已读，false=仅未读
        - in: query
          name: favorite
          schema: { type: boolean }
          description: true=仅收藏
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: page_size
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Article'
                  total:
                    type: integer
                    format: int64

  /articles/{id}/read:
    put:
      tags: [articles]
      summary: 标记文章为已读
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: 文章 ID
      responses:
        '200':
          description: 标记成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/Error'

  /articles/{id}/ai/translate:
    post:
      tags: [articles]
      summary: 手动翻译文章
      description: |
        对单篇文章执行同步 AI 翻译。若文章尚无 AI 分类，会在同一次模型调用中同时生成分类与译文。
        `ai_model_id` 与 `ai_target_language` 都是可选字段：
        - 不传 `ai_model_id` 时，使用该文章所属订阅上配置的默认 AI 模型
        - 不传 `ai_target_language` 或传空串时，使用订阅上配置的默认翻译目标语言
        - 若文章已有译文，接口会返回错误
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: 文章 ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualArticleTranslateRequest'
      responses:
        '200':
          description: 翻译成功，返回最新文章内容
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleEnvelope'
        '400':
          description: 参数错误、缺少模型/目标语言、文章已有译文、AI 正在处理中，或模型调用失败
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIErrorBody'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 文章不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIErrorBody'
        '503':
          description: AI 服务不可用
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /articles/{id}/ai/translate/stream:
    post:
      tags: [articles]
      summary: 手动翻译文章（流式）
      description: |
        对单篇文章执行流式 AI 翻译（SSE）。若文章尚无 AI 分类，会在同一次模型调用中同时生成分类与译文；`delta` 事件仍只输出译文正文增量。
        请求参数与 `/articles/{id}/ai/translate` 相同：
        - 不传 `ai_model_id` 时，使用该文章所属订阅上配置的默认 AI 模型
        - 不传 `ai_target_language` 或传空串时，使用订阅上配置的默认翻译目标语言
        - 若文章已有译文，接口会返回错误事件

        成功时响应 `text/event-stream`，事件统一通过 `data:` 发送 JSON：
        - `{"delta":"..."}`：译文正文增量（前端应持续拼接并渲染）
        - `{"article":{...}}`：翻译完成，返回最终文章对象
        - `{"error":"...","ai_last_error":"..."}`：流内错误
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
          description: 文章 ID
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ManualArticleTranslateRequest'
      responses:
        '200':
          description: |
            流式输出（SSE）。首部为 `Content-Type: text/event-stream`。
            前端需按行解析 `data:` 事件。
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                delta:
                  summary: 增量事件
                  value: |
                    data: {"delta":"<p>第一段译文"}
                done:
                  summary: 完成事件
                  value: |
                    data: {"article":{"id":123,"title_translated":"...","content_translated":"..."}}
                error:
                  summary: 错误事件
                  value: |
                    data: {"error":"AI 处理失败","ai_last_error":"模型超时"}
        '400':
          description: 参数错误（仅在进入 SSE 之前返回 JSON）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIErrorBody'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: 文章不存在（仅在进入 SSE 之前返回 JSON）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIErrorBody'
        '503':
          description: AI 服务不可用（仅在进入 SSE 之前返回 JSON）
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'

  /summary-histories:
    get:
      tags: [summary-histories]
      summary: 总结历史列表
      security: [{ bearerAuth: [] }]
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SummaryHistoryItem'
                  total:
                    type: integer
                    format: int64

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Unauthorized:
      description: 未认证或 Token 无效
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
    Error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'

  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: string

    AIErrorBody:
      type: object
      properties:
        error:
          type: string
        ai_last_error:
          type: string
          description: 当服务端已记录最近一次模型失败原因时返回

    LoginResult:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ArticleEnvelope:
      type: object
      properties:
        article:
          $ref: '#/components/schemas/Article'

    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        status: { type: string, description: active | locked }
        is_super_admin: { type: boolean }
        feishu_id: { type: string, nullable: true }
        feishu_name: { type: string, nullable: true }
        created_at: { type: string, format: date-time }

    FeedCategory:
      type: object
      properties:
        id: { type: integer }
        user_id: { type: integer }
        name: { type: string, maxLength: 64 }
        sort_order: { type: integer }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    CategoryUpsertRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
          description: 分类名称，服务端会自动去除首尾空白

    CategoryReorderRequest:
      type: object
      required: [id_list]
      properties:
        id_list:
          type: array
          description: 当前用户分类 ID 的完整有序列表
          items:
            type: integer

    Feed:
      type: object
      description: 字段与 web 一致，此处列出常用项
      properties:
        id: { type: integer }
        user_id: { type: integer }
        category_id: { type: integer, nullable: true }
        url: { type: string }
        title: { type: string }
        update_interval_minutes: { type: integer }
        expire_days: { type: integer }
        ai_model_id: { type: integer, nullable: true }
        ai_classify_enabled: { type: boolean }
        ai_translate_enabled: { type: boolean }
        ai_target_language: { type: string }
        last_fetched_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }

    Article:
      type: object
      properties:
        id: { type: integer }
        feed_id: { type: integer }
        guid: { type: string }
        title: { type: string }
        link: { type: string }
        content: { type: string }
        ai_process_status: { type: string }
        ai_last_error: { type: string }
        ai_category: { type: string }
        ai_category_translated: { type: string }
        title_translated: { type: string }
        content_translated:
          type: string
          description: 译文正文；新翻译结果会尽量保留原始 HTML 结构与媒体/嵌入节点，旧数据也可能仍是纯文本
        published_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        read: { type: boolean }
        favorite: { type: boolean }
        feed_title: { type: string }
        feed_ai_translate_enabled: { type: boolean }
        feed_ai_classify_enabled: { type: boolean }
        feed_ai_model_id: { type: integer, nullable: true }
        feed_ai_target_language: { type: string }

    ManualArticleTranslateRequest:
      type: object
      properties:
        ai_model_id:
          type: integer
          nullable: true
          description: 可选，临时覆盖订阅默认 AI 模型
        ai_target_language:
          type: string
          description: |
            可选，临时覆盖订阅默认翻译目标语言。
            允许值与服务端一致：`zh-CN`、`en`、`fr`、`de`、`ar`

    SummaryHistoryItem:
      type: object
      properties:
        id: { type: integer }
        ai_model_id: { type: integer }
        ai_model_name: { type: string }
        summary_template_id: { type: integer, nullable: true }
        summary_template_name: { type: string }
        start_time: { type: string }
        end_time: { type: string }
        page: { type: integer }
        page_size: { type: integer }
        order: { type: string }
        article_count: { type: integer }
        total: { type: integer, format: int64 }
        content: { type: string }
        error: { type: string }
        created_at: { type: string, format: date-time }